Enemy update :bowtie:
+Forest watcher now have spritesheet (Needs tuning though)
+Stone crawler prefab (no death animation)(Needs AI)
+4 legged moss prefab (no death animation)
+Level 1 and 2 has their unique monsters

The spawning should be somewhat balanced. However the stone crawler
cannot be killed as of now. It needs hide AI. Also I couldn't figure out
how to put it exactly on top of the ground? (This should be the only
monster to have that since it's supposed to blend in the background.)

Bug report: When 4 legged moss is active at goal game crashes. (And so
do unity fml) I thought it was that 4 legged moss missed enemy tag but
it doesn't. I'm calling for you hackerman!